Prevents crash on empty area filter
Previously, passing an empty or undefined areafilter in the request options from the frontend lead to a crash. The server code tried adding the first point from the area filter again to close the geometry shape, but this failed since there were no points in the filter at all.

This adds an extra check to only run the filter logic if there are > 0
points in the area filter.